requirements.txt
torch
torchvision
numpy
pandas
tqdm
pillow
opencv-python
scikit-learn
open_clip_torch
adversarial-robustness-toolbox
pyyaml
